docker-services/essentiels/qrcodes/requirements.txt
flask==2.3.3
flask-cors==4.0.0
qrcode[pil]==7.4.2
pillow==10.0.1
psycopg2-binary==2.9.7
pyjwt==2.8.0
requests==2.31.0
